[ring buffer] reduced ring buffer size to avoid overflow of rom memory

+++ ring_buffer.c
@@ -6,7 +6,7 @@ void ring_buffer_init(struct ring_buffer *rb) {
     rb->write_pointer = 0;
     rb->available_bytes = 0;
 
-    for (uint16_t i = 0; i < RING_BUFFER_SIZE; i++) {
+    for (uint8_t i = 0; i < RING_BUFFER_SIZE; i++) {
         rb->buffer[i] = 0;
     }
 }
@@ -32,7 +32,7 @@ uint8_t ring_buffer_get(struct ring_buffer *rb) {
     return return_value;
 }
 
-uint16_t ring_buffer_available_bytes(struct ring_buffer *rb) {
+uint8_t ring_buffer_available_bytes(struct ring_buffer *rb) {
     return rb->available_bytes;
 }
